init
Contracts code version same as gh commit 55096cf1fd091629ff8dad783f71fb4758eded46 on original repo.

You can verify that code inside ./contracts folder is the one that was audited, you can compare the hashes of each file in the ./contracts folder with the hash that is in the Quantstamp certificate.